update tests, move Upcoming Departures into file
