+++ apps/site/assets/ts/schedule/__tests__/SchedulePageTest.tsx
@@ -4,7 +4,7 @@ import { createReactRoot } from "../../app/helpers/testUtils";
 import SchedulePage from "../components/SchedulePage";
 import { TypedRoutes } from "../../stop/components/__stop";
 import ScheduleNote from "../components/ScheduleNote";
-import { EnhancedRoute, Service } from "../../__v3api";
+import { EnhancedRoute, Service, ServiceWithServiceDate } from "../../__v3api";
 
 const pdfs = [
   {
@@ -58,6 +58,7 @@ const route: EnhancedRoute = {
   long_name: "Orange Line",
   type: 1
 };
+
 const stops = [
   {
     name: "Malden Center",
@@ -73,7 +74,7 @@ const stops = [
   }
 ];
 
-const service: Service = {
+const service: ServiceWithServiceDate = {
   added_dates: [],
   added_dates_notes: {},
   description: "Weekday schedule",
@@ -84,7 +85,8 @@ const service: Service = {
   start_date: "2019-06-25",
   type: "weekday",
   typicality: "typical_service",
-  valid_days: [1, 2, 3, 4, 5]
+  valid_days: [1, 2, 3, 4, 5],
+  service_date: "2019-06-26"
 };
 
 const services = [service];
